requirements.txt
streamlit
google-generativeai
markdown-pdf